Tag 2.1.1

+++ classes/base-deps-resolver.bbclass
@@ -341,10 +341,10 @@ def sls_generate_native_sysroot(d):
     bb.build.exec_func("sysroot_stage_all", d)
     fixme_path = d.expand("${SYSROOT_DESTDIR}${base_prefix}/")
     fixme_file_path = os.path.join(sysroot_components_dir,"fixmepath")
-    if os.path.exists(fixme_file_path):
+    if os.path.exists(fixme_file_path) and os.path.exists(fixme_path):
         shutil.copy(fixme_file_path,fixme_path)
     ver_file_path = os.path.join(sysroot_components_dir,"version-%s"%d.getVar("PN"))
-    if os.path.exists(ver_file_path):
+    if os.path.exists(ver_file_path) and os.path.exists(fixme_path):
         shutil.copy(ver_file_path,fixme_path)
     pn = d.getVar("PN")
     multiprov = d.getVar("BB_MULTI_PROVIDER_ALLOWED").split()
